update plugins one

+++ wp-content/plugins/ajax-load-more-pro/ajax-load-more-pro.php
@@ -6,9 +6,10 @@
  * Author: Darren Cooney
  * Twitter: @KaptonKaos
  * Author URI: https://connekthq.com
- * Version: 1.3.0.1
+ * Version: 1.3.2.1
  * License: GPL
  * Copyright: Darren Cooney & Connekt Media
+ * Requires Plugins: ajax-load-more
  *
  * @package ALMPro
  */
@@ -17,8 +18,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-define( 'ALM_PRO_VERSION', '1.3.0.1' );
-define( 'ALM_PRO_RELEASE', 'January 22, 2025' );
+define( 'ALM_PRO_VERSION', '1.3.2.1' );
+define( 'ALM_PRO_RELEASE', 'June 10, 2025' );
 
 /**
  * Plugin installation hook
@@ -35,17 +36,17 @@ function alm_pro_install() {
 			return false;
 		}
 
-		// Loop all addons.
-		foreach ( $ajax_load_more->alm_return_addons() as $plugin ) {
+		if ( method_exists( $ajax_load_more, 'alm_return_addons' ) ) {
+			foreach ( $ajax_load_more->alm_return_addons() as $plugin ) {
+				// Check if standalone addon is active.
+				if ( is_plugin_active( $plugin['path'] . '/' . $plugin['path'] . '.php' ) ) {
+					deactivate_plugins( $plugin['path'] . '/' . $plugin['path'] . '.php' ); // deactivate it.
+				}
 
-			// Check if standalone addon is active.
-			if ( is_plugin_active( $plugin['path'] . '/' . $plugin['path'] . '.php' ) ) {
-				deactivate_plugins( $plugin['path'] . '/' . $plugin['path'] . '.php' ); // deactivate it.
-			}
-
-			// Set status option.
-			if ( ! get_option( 'alm_pro_status_' . $plugin['slug'] ) ) {
-				update_option( 'alm_pro_status_' . $plugin['slug'], 'active' );
+				// Set status option.
+				if ( ! get_option( 'alm_pro_status_' . $plugin['slug'] ) ) {
+					update_option( 'alm_pro_status_' . $plugin['slug'], 'active' );
+				}
 			}
 		}
 	}
@@ -58,7 +59,7 @@ register_activation_hook( __FILE__, 'alm_pro_install' );
  * @since 1.2.19
  */
 function alm_pro_admin_notice() {
-	$slug   = 'ajax-load-more';
+	$slug = 'ajax-load-more';
 	// Ajax Load More Notice.
 	if ( get_transient( 'alm_pro_admin_notice' ) ) {
 		$install_url = get_admin_url() . '/update.php?action=install-plugin&plugin=' . $slug . '&_wpnonce=' . wp_create_nonce( 'install-plugin_' . $slug );
@@ -83,11 +84,20 @@ if ( ! class_exists( 'ALMPro' ) ) :
 		 * Construct Class.
 		 */
 		public function __construct() {
-			add_action( 'alm_pro_installed', array( &$this, 'alm_pro_installed' ) );
-			add_action( 'plugins_loaded', array( &$this, 'alm_pro_load_addons' ) );
-			add_action( 'wp_ajax_alm_pro_toggle_activation', array( &$this, 'alm_pro_toggle_activation' ) );
-			load_plugin_textdomain( 'ajax-load-more-pro', false, dirname( plugin_basename( __FILE__ ) ) . '/lang/' );
+			add_action( 'alm_pro_installed', [ &$this, 'alm_pro_installed' ] );
+			add_action( 'plugins_loaded', [ &$this, 'alm_pro_load_addons' ] );
+			add_action( 'wp_ajax_alm_pro_toggle_activation', [ &$this, 'alm_pro_toggle_activation' ] );
+			add_action( 'init', [ $this, 'alm_pro_textdomain' ] );
 			$this->constants();
+		}
+
+		/**
+		 * Load text domain.
+		 *
+		 * @return void
+		 */
+		public function alm_pro_textdomain() {
+			load_plugin_textdomain( 'ajax-load-more-pro', false, dirname( plugin_basename( __FILE__ ) ) . '/lang/' );
 		}
 
 		/**
@@ -107,17 +117,17 @@ if ( ! class_exists( 'ALMPro' ) ) :
 		 * @since 1.0
 		 */
 		public function alm_pro_load_addons() {
-
 			global $ajax_load_more;
 			if ( ! $ajax_load_more ) {
 				return false;
 			}
 
-			// Loop addons.
-			foreach ( $ajax_load_more->alm_return_addons() as $plugin ) {
-				if ( ! has_action( $plugin['action'] ) ) {
-					if ( 'active' === get_option( ALM_PRO_OPTION_PREFIX . $plugin['slug'] ) ) {
-						require_once 'pro/' . $plugin['path'] . '/' . $plugin['path'] . '.php';
+			if ( method_exists( $ajax_load_more, 'alm_return_addons' ) ) {
+				foreach ( $ajax_load_more->alm_return_addons() as $plugin ) {
+					if ( ! has_action( $plugin['action'] ) ) {
+						if ( 'active' === get_option( ALM_PRO_OPTION_PREFIX . $plugin['slug'] ) ) {
+							require_once 'pro/' . $plugin['path'] . '/' . $plugin['path'] . '.php';
+						}
 					}
 				}
 			}
@@ -129,7 +139,7 @@ if ( ! class_exists( 'ALMPro' ) ) :
 		 * @since 1.0
 		 */
 		public static function alm_enqueue_pro_admin_scripts() {
-			wp_enqueue_script( 'alm-pro-admin', ALM_PRO_ADMIN_URL . '/admin/js/ajax-load-more-pro.js', array( 'jquery' ), ALM_PRO_VERSION, false );
+			wp_enqueue_script( 'alm-pro-admin', ALM_PRO_ADMIN_URL . '/admin/js/ajax-load-more-pro.js', [ 'jquery' ], ALM_PRO_VERSION, false );
 		}
 
 		/**
@@ -142,7 +152,6 @@ if ( ! class_exists( 'ALMPro' ) ) :
 			$slug  = $_POST['slug'];
 
 			if ( $slug && current_user_can( apply_filters( 'alm_custom_user_role', 'edit_theme_options' ) ) ) {
-
 				// Check the nonce, don't match then bounce!
 				if ( ! wp_verify_nonce( $nonce, 'alm_repeater_nonce' ) ) {
 					die( esc_html__( 'Error - Unable to verify nonce.', 'ajax-load-more-pro' ) );
@@ -156,21 +165,21 @@ if ( ! class_exists( 'ALMPro' ) ) :
 					update_option( ALM_PRO_OPTION_PREFIX . $slug, $result );
 				}
 
-				$return = array(
+				$return = [
 					'success' => true,
 					'slug'    => $slug,
 					'result'  => $result,
 					'msg'     => esc_html__( 'Add-on status updated', 'ajax-load-more-pro' ),
-				);
+				];
 				wp_send_json( $return );
 
 			} else {
-				$return = array(
+				$return = [
 					'success' => false,
 					'slug'    => $slug,
 					'result'  => 'error',
 					'msg'     => esc_html__( 'Add-on status NOT updated', 'ajax-load-more-pro' ),
-				);
+				];
 				wp_send_json( $return );
 
 			}
@@ -248,12 +257,12 @@ function alm_pro_plugin_updater() {
 	$edd_updater = new EDD_SL_Plugin_Updater(
 		ALM_STORE_URL,
 		__FILE__,
-		array(
+		[
 			'version' => ALM_PRO_VERSION,
 			'license' => $license_key,
 			'item_id' => ALM_PRO_ITEM_NAME,
 			'author'  => 'Darren Cooney',
-		)
+		]
 	);
 }
 add_action( 'admin_init', 'alm_pro_plugin_updater', 0 );
